pyproject.toml
[tool.poetry]
name = "discord-token-abuse-killer"
version = "0.1.0"
description = ""
authors = ["Jack Webb <git@jackwebb.me>"]

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
yarl = "^1.7.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
